[chore] Add ability to configure sizer for default batcher

+++ exporter/exporterhelper/internal/queue_sender.go
@@ -203,9 +203,9 @@ type SizeConfig struct {
 	Sizer request.SizerType `mapstructure:"sizer"`
 
 	// MinSize defines the configuration for the minimum size of a batch.
-	MinSize int `mapstructure:"min_size"`
+	MinSize int64 `mapstructure:"min_size"`
 	// MaxSize defines the configuration for the maximum size of a batch.
-	MaxSize int `mapstructure:"max_size"`
+	MaxSize int64 `mapstructure:"max_size"`
 }
 
 func (c *BatcherConfig) Validate() error {
